change package struct
