+++ pom.xml
@@ -5,7 +5,7 @@
 
     <groupId>xws</groupId>
     <artifactId>neuron</artifactId>
-    <version>0.0.1</version>
+    <version>0.0.2</version>
     <packaging>jar</packaging>
 
     <name>xws-neuron</name>
