+++ src/main/java/xws/neuron/graph/Operation.java
@@ -1,0 +1,10 @@
+package xws.neuron.graph;
+
+/**
+ * the operation of graph
+ * Created by xws on 2019/9/1.
+ */
+public class Operation {
+
+
+}
