+++ src/main/java/xws/neuron/layer/conv/ConvolutionLayer.java
@@ -1,18 +1,18 @@
-package xws.neuron.layer;
+package xws.neuron.layer.conv;
 
 import xws.neuron.ActivationFunction;
 import xws.neuron.CNNFilter;
 import xws.neuron.Tensor;
 import xws.neuron.UtilNeuralNet;
+import xws.neuron.layer.Layer;
 import xws.util.UtilFile;
 
 
 /**
- * 标准反卷积层
- * 输入卷积后的结果，对输入进行还原
+ * 标准卷积层
  * Created by xws on 2019/2/19.
  */
-public class DeconvolutionLayer extends Layer {
+public class ConvolutionLayer extends Layer {
 
 
     private CNNFilter[] filters;//卷积核
@@ -49,12 +49,12 @@ public class DeconvolutionLayer extends Layer {
     //z值数据存储
     private Tensor tensorZ;
 
-    public DeconvolutionLayer() {
+    public ConvolutionLayer() {
     }
 
     //构造函数时，传入filters的构造,total是特征
-    public DeconvolutionLayer(int total, int height, int width, int strideX, int strideY, int padding) {
-        super("filter");
+    public ConvolutionLayer(int total, int height, int width, int strideX, int strideY, int padding) {
+        super("ConvolutionLayer");
         this.height = height;
         this.width = width;
         this.strideX = strideX;
@@ -65,8 +65,8 @@ public class DeconvolutionLayer extends Layer {
     }
 
 
-    public DeconvolutionLayer(String name, String activationType, int total, int height, int width, int strideX, int strideY, int padding) {
-        super("filter");
+    public ConvolutionLayer(String name, String activationType, int total, int height, int width, int strideX, int strideY, int padding) {
+        super("ConvolutionLayer");
         setName(name);
         setActivationType(activationType);
         this.height = height;
@@ -78,8 +78,8 @@ public class DeconvolutionLayer extends Layer {
 
     }
 
-    public DeconvolutionLayer(String name, String activationType, int total, int height, int width, int strideX, int strideY, int padding, double lambda) {
-        super("filter");
+    public ConvolutionLayer(String name, String activationType, int total, int height, int width, int strideX, int strideY, int padding, double lambda) {
+        super("ConvolutionLayer");
         setName(name);
         setActivationType(activationType);
         this.height = height;
@@ -101,12 +101,18 @@ public class DeconvolutionLayer extends Layer {
 
         initFile();
 
-        //反向卷积，传递正向卷积结果进来
-        //首先，计算反卷积后的大小，并开辟空间
         tensorInput = tensor;
+
+        //如果卷积层前面是FullLayer,需要计算出结果数据的维度和大小
+        if (tensor.getHeight() == 1) {
+            int val = (int) Math.sqrt((double) tensor.getArray().length);
+            tensor.setHeight(val);
+            tensor.setWidth(val);
+        }
+
         outDepth = filters.length;
-        outHeight = (tensor.getHeight() - 1) * strideY + height;
-        outWidth = (tensor.getWidth() - 1) * strideX + width;
+        outHeight = UtilNeuralNet.afterHeight(tensor.getHeight(), padding, strideY, height);
+        outWidth = UtilNeuralNet.afterWidth(tensor.getWidth(), padding, strideX, width);
 
 
         Tensor tensorOut = new Tensor();
@@ -135,7 +141,8 @@ public class DeconvolutionLayer extends Layer {
             //有多少输出，意味着就要执行多少次卷积操作
             for (int h = 0; h < outHeight; h++) {
                 for (int w = 0; w < outWidth; w++) {
-                    cnnFilter.deconvolution(d, h * strideY, w * strideX, tensorInput, tensorOut);
+                    double val = cnnFilter.convolution(h * strideY, w * strideX, tensorInput);
+                    tensorOut.set(d, h, w, val);
                 }
             }
         }
@@ -163,11 +170,11 @@ public class DeconvolutionLayer extends Layer {
     public Tensor backPropagation(Tensor tensor) {
 
         //创建张量，存储误差数据
-        Tensor tensorErrorInput = new Tensor();
-        tensorErrorInput.setDepth(tensorInput.getDepth());
-        tensorErrorInput.setHeight(tensorInput.getHeight());
-        tensorErrorInput.setWidth(tensorInput.getWidth());
-        tensorErrorInput.createArray();
+        Tensor tensorError = new Tensor();
+        tensorError.setDepth(tensorInput.getDepth());
+        tensorError.setHeight(tensorInput.getHeight());
+        tensorError.setWidth(tensorInput.getWidth());
+        tensorError.createArray();
 
         pdz(tensor);
 
@@ -181,8 +188,8 @@ public class DeconvolutionLayer extends Layer {
             for (int h = 0; h < outHeight; h++) {
                 for (int w = 0; w < outWidth; w++) {
                     //获取误差
-                    double error = filter.deconvolution_d(d, h * strideY, w * strideX, tensor, tensorErrorInput);
-                    tensorErrorInput.set(d, h, w, error);
+                    double error = tensor.get(d, h, w);
+                    filter.convolution_d(h * strideY, w * strideX, tensorError, error);
                 }
             }
         }
@@ -209,7 +216,7 @@ public class DeconvolutionLayer extends Layer {
             filter.updateErrorB(tensor, i, getLearnRate());
         }
 
-        return tensorErrorInput;
+        return tensorError;
     }
 
     //求∂A/∂Z
