+++ src/main/java/xws/neuron/layer/conv/DepthSeparableLayer.java
@@ -1,9 +1,10 @@
-package xws.neuron.layer;
+package xws.neuron.layer.conv;
 
 import xws.neuron.ActivationFunction;
 import xws.neuron.CNNFilter;
 import xws.neuron.Tensor;
 import xws.neuron.UtilNeuralNet;
+import xws.neuron.layer.Layer;
 import xws.util.UtilFile;
