+++ src/main/java/xws/test/MnistTest.java
@@ -5,6 +5,9 @@ import xws.neuron.CNNetWork;
 import xws.neuron.Tensor;
 import xws.neuron.UtilNeuralNet;
 import xws.neuron.layer.*;
+import xws.neuron.layer.bn.BnLayer;
+import xws.neuron.layer.conv.ConvolutionLayer;
+import xws.neuron.layer.pool.MaxPoolLayer;
 import xws.util.Cifar10;
 import xws.util.UtilCifar10;
 import xws.util.UtilMnist;
