+++ src/main/java/xws/test/PoolTest.java
@@ -1,8 +1,8 @@
 package xws.test;
 
 import xws.neuron.Tensor;
-import xws.neuron.layer.MaxPoolLayer;
-import xws.neuron.layer.MeanPoolLayer;
+import xws.neuron.layer.pool.MaxPoolLayer;
+import xws.neuron.layer.pool.MeanPoolLayer;
 import xws.neuron.layer.PaddingLayer;
 
 /**
